fix: parse close frame & move failWebsocketConnection
