+++ lib/websocket/connection.js
@@ -9,12 +9,10 @@ const {
   kResponse,
   kExtensions,
   kProtocol,
-  kClosingFrame,
   kSentClose,
-  kByteParser,
-  kController
+  kByteParser
 } = require('./symbols')
-const { fireEvent, isEstablished } = require('./util')
+const { fireEvent, isEstablished, failWebsocketConnection } = require('./util')
 const { CloseEvent } = require('./events')
 const { ByteParser } = require('./receiver')
 const { makeRequest } = require('../fetch/request')
@@ -197,19 +195,6 @@ function establishWebSocketConnection (url, protocols, ws) {
 }
 
 /**
- * @param {import('./websocket').WebSocket} ws
- */
-function failWebsocketConnection (ws) {
-  const { [kController]: controller, [kResponse]: response } = ws
-
-  controller.abort()
-
-  if (!response.socket.destroyed) {
-    response.socket.destroy()
-  }
-}
-
-/**
  * @see https://websockets.spec.whatwg.org/#feedback-from-the-protocol
  * @param {import('./websocket').WebSocket} ws
  */
@@ -276,14 +261,9 @@ function socketClosed (ws) {
     let code = 1005
     let reason = ''
 
-    if (ws[kClosingFrame]) {
-      /** @type {ReturnType<WebsocketFrame['parseCloseBody']} */
-      const result = ws[kClosingFrame].parseCloseBody()
+    const result = ws[kByteParser].closingInfo
 
-      if (result === null) {
-        return
-      }
-
+    if (result) {
       code = result.code ?? 1005
       reason = result.reason
     } else if (!ws[kSentClose]) {
@@ -327,6 +307,5 @@ function socketClosed (ws) {
 }
 
 module.exports = {
-  establishWebSocketConnection,
-  failWebsocketConnection
+  establishWebSocketConnection
 }
