+++ lib/websocket/receiver.js
@@ -1,8 +1,7 @@
 const { Writable } = require('stream')
 const { parserStates, opcodes, states } = require('./constants')
-const { failWebsocketConnection } = require('./connection')
-const { kReadyState } = require('./symbols')
-const { isValidStatusCode } = require('./util')
+const { kReadyState, kSentClose } = require('./symbols')
+const { isValidStatusCode, failWebsocketConnection } = require('./util')
 
 class ByteParser extends Writable {
   #buffers = []
@@ -80,10 +79,17 @@ class ByteParser extends Writable {
           return
         }
 
+        const body = buffer.subarray(2, payloadLength)
+
+        this.#info.closeInfo = this.parseCloseBody(true, body)
+
         // Upon either sending or receiving a Close control frame, it is said
         // that _The WebSocket Closing Handshake is Started_ and that the
         // WebSocket connection is in the CLOSING state.
-        this.ws[kReadyState] = states.CLOSED
+        this.ws[kReadyState] = states.CLOSING
+        this.ws[kSentClose] = true
+
+        return
       }
 
       // TODO: handle control frames here. Since they are unfragmented, and can
@@ -149,12 +155,10 @@ class ByteParser extends Writable {
     }
   }
 
-  parseCloseBody (onlyCode) {
+  parseCloseBody (onlyCode, data) {
     // https://datatracker.ietf.org/doc/html/rfc6455#section-7.1.5
     /** @type {number|undefined} */
     let code
-
-    const { data } = this.#info
 
     if (data.length > 2) {
       // _The WebSocket Connection Close Code_ is
@@ -192,6 +196,10 @@ class ByteParser extends Writable {
     }
 
     return { code, reason }
+  }
+
+  get closingInfo () {
+    return this.#info.closeInfo
   }
 }
