+++ lib/websocket/util.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { kReadyState } = require('./symbols')
+const { kReadyState, kController, kResponse } = require('./symbols')
 const { states } = require('./constants')
 
 /**
@@ -118,11 +118,25 @@ function isValidStatusCode (code) {
   return code >= 3000 && code <= 4999
 }
 
+/**
+ * @param {import('./websocket').WebSocket} ws
+ */
+function failWebsocketConnection (ws) {
+  const { [kController]: controller, [kResponse]: response } = ws
+
+  controller.abort()
+
+  if (response?.socket && !response.socket.destroyed) {
+    response.socket.destroy()
+  }
+}
+
 module.exports = {
   isEstablished,
   isClosing,
   isClosed,
   fireEvent,
   isValidSubprotocol,
-  isValidStatusCode
+  isValidStatusCode,
+  failWebsocketConnection
 }
