+++ lib/websocket/websocket.js
@@ -14,8 +14,8 @@ const {
   kResponse,
   kSentClose
 } = require('./symbols')
-const { isEstablished, isClosing, isValidSubprotocol } = require('./util')
-const { establishWebSocketConnection, failWebsocketConnection } = require('./connection')
+const { isEstablished, isClosing, isValidSubprotocol, failWebsocketConnection } = require('./util')
+const { establishWebSocketConnection } = require('./connection')
 const { WebsocketFrameSend } = require('./frame')
 const { kEnumerableProperty, isBlobLike } = require('../core/util')
 const { types } = require('util')
